+++ site/dat/docs/Changelog.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 1.4.0 (next)
+ - add XPath extractors and assertions for JMeter
+ - show warning if no element matched `set-prop` modification
+ - do not create hostaliases file when aliases are not specified
+ - add `force-parent-sample` option to JMeter executor
+ - add `compile-target-java` option for Selenium
+
+## 1.3.3 <sup>24 mar 2016</sup>
+ - add new `hostaliases` setting for all executors
+ - add delay capability to engine 
+
 ## 1.3.2 <sup>23 mar 2016</sup>
  - fix lowercase hostname for JMeter HTTP request
  - fix binary varname crash for JMeter HTTP request